Phase 3: Move 3 serverList MergedFollow*State files to commons
Move MergedFollowPlusMineRelayListsState, MergedFollowPlusMineWithIndex-
RelayListsState, MergedFollowPlusMineWithSearchRelayListsState to commons.
MergedFollowListsState skipped (depends on FollowListsState/PeopleListsState
not yet in commons). Leave typealiases at original locations.
