+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/serverList/MergedFollowPlusMineRelayListsState.kt
@@ -20,63 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.serverList
 
-import com.vitorpamplona.amethyst.model.edits.PrivateStorageRelayListState
-import com.vitorpamplona.amethyst.model.localRelays.LocalRelayListState
-import com.vitorpamplona.amethyst.model.nip02FollowLists.FollowListOutboxOrProxyRelays
-import com.vitorpamplona.amethyst.model.nip65RelayList.Nip65RelayListState
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-
-class MergedFollowPlusMineRelayListsState(
-    val followsOutboxOrProxyRelayList: FollowListOutboxOrProxyRelays,
-    val nip65RelayList: Nip65RelayListState,
-    val privateOutboxRelayList: PrivateStorageRelayListState,
-    val localRelayList: LocalRelayListState,
-    val scope: CoroutineScope,
-) {
-    fun mergeLists(lists: Array<Set<NormalizedRelayUrl>>): Set<NormalizedRelayUrl> = lists.reduce { acc, set -> acc + set }
-
-    val flow: StateFlow<Set<NormalizedRelayUrl>> =
-        combine(
-            listOf(
-                followsOutboxOrProxyRelayList.flow,
-                nip65RelayList.outboxFlow,
-                nip65RelayList.inboxFlow,
-                privateOutboxRelayList.flow,
-                localRelayList.flow,
-            ),
-            ::mergeLists,
-        ).onStart {
-            emit(
-                mergeLists(
-                    arrayOf(
-                        followsOutboxOrProxyRelayList.flow.value,
-                        nip65RelayList.outboxFlow.value,
-                        nip65RelayList.inboxFlow.value,
-                        privateOutboxRelayList.flow.value,
-                        localRelayList.flow.value,
-                    ),
-                ),
-            )
-        }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                mergeLists(
-                    arrayOf(
-                        followsOutboxOrProxyRelayList.flow.value,
-                        nip65RelayList.outboxFlow.value,
-                        nip65RelayList.inboxFlow.value,
-                        privateOutboxRelayList.flow.value,
-                        localRelayList.flow.value,
-                    ),
-                ),
-            )
-}
+typealias MergedFollowPlusMineRelayListsState = com.vitorpamplona.amethyst.commons.model.serverList.MergedFollowPlusMineRelayListsState
